remove synchronous routing
