+++ Sources/Async/Futures/Promise+Helpers.swift
@@ -4,3 +4,9 @@ extension Promise {
         future.then(callback: self.complete).catch(callback: self.fail)
     }
 }
+
+extension FutureType where Expectation == Self {
+    public func addAwaiter(callback: @escaping ((FutureResult<Self>) -> ())) {
+        callback(.expectation(self))
+    }
+}
