+++ Sources/Development/main.swift
@@ -28,21 +28,20 @@ services.register { container in
 
 let app = Application(services: services)
 
-let async = try app.make(AsyncRouter.self)
-let sync = try app.make(SyncRouter.self)
+let router = try app.make(Router.self)
 
 let user = User(name: "Vapor", age: 3);
-async.get("hello") { req in
+router.get("hello") { req in
     return Future<User>(user)
 }
 
 let hello = try Response(body: "Hello, world!")
-sync.get("plaintext") { req in
+router.get("plaintext") { req in
     return hello
 }
 
 let view = try app.make(ViewRenderer.self)
-async.get("leaf") { req -> Future<View> in
+router.get("leaf") { req -> Future<View> in
     user.child = User(name: "Leaf", age: 1)
     let promise = Promise(User.self)
     user.futureChild = promise.future
@@ -78,7 +77,7 @@ extension Worker {
 
 let database = SQLite.Database(path: "/tmp/db.sqlite")
 
-async.get("sqlite") { req -> Future<String> in
+router.get("sqlite") { req -> Future<String> in
     let promise = Promise(String.self)
     
     let pool = try req.requireWorker()
