+++ Sources/Vapor/Async+Helpers.swift
@@ -1,0 +1,7 @@
+import Async
+import HTTP
+
+/// Makes `Response` a drop-in replacement for `Future<Response>
+extension Response: FutureType {
+    public typealias Expectation = Response
+}
