+++ Sources/Vapor/RoutingGroups.swift
@@ -8,7 +8,7 @@ import Routing
 /// All path components will be inserted before the Route's path
 ///
 /// All middleware will be applied to the Responder
-public final class Group : Router, SyncRouter, AsyncRouter {
+public final class Group : Router {
     /// All routes registered to this group
     public private(set) var routes: [Route] = []
